Reorganization of modules.
