+++ django_workflow_system/models/collections/collection_image_type.py
