+++ django_workflow_system/models/collections/collection_member.py
@@ -4,7 +4,7 @@ import uuid
 from django.core.validators import MinValueValidator
 from django.db import models
 
-from django_workflow_system.models.collection import WorkflowCollection
+from django_workflow_system.models.collections.collection import WorkflowCollection
 from django_workflow_system.models.workflow import Workflow
 from django_workflow_system.models.abstract_models import CreatedModifiedAbstractModel
