+++ django_workflow_system/models/step_dependency_detail.py
@@ -5,7 +5,9 @@ from django.core.exceptions import ValidationError, ObjectDoesNotExist
 from django.db import models
 from jsonschema import Draft7Validator, SchemaError
 
-from django_workflow_system.models.collection_member import WorkflowCollectionMember
+from django_workflow_system.models.collections.collection_member import (
+    WorkflowCollectionMember,
+)
 from django_workflow_system.models.step import WorkflowStep
 from django_workflow_system.models.step_dependency_group import (
     WorkflowStepDependencyGroup,
